+++ Migrators/AllureExporter/Services/IAttributeService.cs
@@ -4,5 +4,5 @@ namespace AllureExporter.Services;
 
 public interface IAttributeService
 {
-    Task<List<Attribute>> GetCustomAttributes(int projectId);
+    Task<List<Attribute>> GetCustomAttributes(long projectId);
 }
